pyproject.toml
[tool.poetry]
name = "ml_ops"
version = "0.1.0"
description = ""
authors = ["Andrew <andrew.tokarevv@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">3.10, <3.13"
scikit-learn = "^1.3.1"
matplotlib = "^3.8.0"
pandas = "^2.1.1"
torch = "^2.1.0"
torchvision = "^0.16.0"
pytorch-lightning = "^2.1.2"
dvc = "^3.30.1"
dvc-gdrive = "^2.20.0"
hydra-core = "^1.3.2"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.4.0"
fire = "^0.5.0"

[tool.black]
line-length = 79
target-version = ["py310", "py311", "py312"]
skip-string-normalization = true

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 79
lines_after_imports = 2

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
